% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAQM-package.R
\docType{package}
\name{BAQM-package}
\alias{BAQM}
\alias{BAQM-package}
\title{BAQM: Analytics Functions for Babson AQM Courses}
\description{
Instructor-developed tools for AQM 1000 and AQM 2000 courses using R in the curriculum.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CPA-wrk/BAQM}
  \item \url{https://cpa-wrk.github.io/BAQM/}
  \item Report bugs at \url{https://github.com/CPA-wrk/BAQM/issues}
}

}
\author{
\strong{Maintainer}: Peter Lert \email{plert@babson.edu} [copyright holder]

}
\keyword{internal}
