% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAQM-package.R
\docType{package}
\name{BAQM-package}
\alias{BAQM}
\alias{BAQM-package}
\title{BAQM: Babson Analytics and Quantitative Methods Tools}
\description{
Instructor-developed tools for Analytics and Quantitative Methods (AQM) courses at Babson College. Included are compact descriptive statistics for data frames and lists, expanded reporting and graphics for linear regressions, and formatted reports for best subsets analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CPA-wrk/BAQM}
  \item \url{https://cpa-wrk.github.io/BAQM/}
  \item Report bugs at \url{https://github.com/CPA-wrk/BAQM/issues}
}

}
\author{
\strong{Maintainer}: Peter Lert \email{plert@babson.edu} [copyright holder]

}
\keyword{internal}
