% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.infl.R
\name{lm_plot.infl}
\alias{lm_plot.infl}
\title{Plot Leverage vs. Fitted Values to Visualize Inflential Observations}
\usage{
lm_plot.infl(
  mdl,
  opt = list(),
  parm = list(),
  df = lm_plot.df(mdl, parm),
  plts = list()
)
}
\arguments{
\item{mdl}{A fitted model object (typically from \code{\link[stats]{lm}}).}

\item{opt}{List of options (not used).}

\item{parm}{List of plotting parameters, usually from \code{lm_plot.parms()}.}

\item{df}{Data frame with augmented model data. Defaults to \code{lm_plot.df(mdl)}.}

\item{plts}{List of ggplot objects to which this plot will be added.}
}
\value{
A list containing:
\itemize{
  \item \code{mdl} Fitted model object,
  \item \code{opt} List of options (unchanged),
  \item \code{parm} Parameter list for plotting,
  \item \code{df} Data frame used for plotting,
  \item \code{plts} List of ggplot objects, including the \code{$infl} element.
}
}
\description{
Creates a plot of leverage values versus the linear fitted values, including an identification of points with a large Cook's distance, to visualize high-leverage and influential observations.
}
\details{
The plot visualizes the calculated leverage of individual data points, defined as the diagonal element of the 'hat' matrix, as a function of the fitted values and implicitly relative to their location in the field of predictor variables, and the threshold value of high leverage is indicated. In addition, Cook's distance can be used to label influential points, along with outlier and regular points.
}
\examples{
\dontrun{
mdl <- lm(Sepal.Length ~ Sepal.Width, data = iris)
result <- lm_plot.infl(mdl)
print(result$plts$infl)
}
}
\seealso{
\code{\link{lm_plot.df}}, \code{\link{lm_plot.parms}}, \code{\link{outlier}}
}
