% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.infl.R
\name{lm_plot.infl}
\alias{lm_plot.infl}
\title{Plot Studentized Residuals vs. Sequence with Influence Identification}
\usage{
lm_plot.infl(mdl, parms = list(), df = lm_plot.df(mdl), plts = list())
}
\arguments{
\item{mdl}{An object of class \code{lm} or similar, representing the fitted model.}

\item{parms}{List of plot customization parameters. Passed to \code{lm_plot.parms()}.}

\item{df}{Data frame with augmented model data. Defaults to \code{lm_plot.df(mdl)}.}

\item{plts}{List of ggplot objects to build upon. Defaults to empty list.}
}
\value{
A list containing:
  \item{mdl}{The model object}
  \item{parms}{Parameters used for plotting}
  \item{df}{Data frame used for plotting}
  \item{plts}{A list of ggplot objects, including the influence plot}
}
\description{
Creates a "comb" plot of studentized residuals versus sequence, highlighting influence and outlier points using a robust outlier heuristic.
}
\details{
The plot visualizes studentized residuals by sequence order, marking regular, outlier, and influential points. Influence thresholds are set using robust outlier detection. Influential and outlier points can be labeled, and influence threshold lines are shown.
}
\examples{
\dontrun{
mdl <- lm(y ~ x, data = mydata)
result <- lm_plot.infl(mdl)
print(result$plts$infl)
}
}
\seealso{
\code{\link{lm_plot.df}}, \code{\link{lm_plot.parms}}, \code{\link{outlier}}
}
