% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.qq.R
\name{lm_plot.qq}
\alias{lm_plot.qq}
\title{Q-Q Plot of Residuals to Assess Normality}
\usage{
lm_plot.qq(mdl, parms = list(), df = lm_plot.df(mdl), plts = list())
}
\arguments{
\item{mdl}{An object of class \code{lm}, representing the fitted linear model.}

\item{parms}{List of plot customization parameters. Passed to \code{lm_plot.parms()}.}

\item{df}{Data frame with augmented model data. Defaults to \code{lm_plot.df(mdl)}.}

\item{plts}{List of ggplot objects to build upon. Defaults to empty list.}
}
\value{
A list containing:
  \item{mdl}{The model object}
  \item{parms}{Parameters used for plotting, including Q-Q line and test results}
  \item{df}{Data frame used for plotting}
  \item{plts}{A list of ggplot objects, including the Q-Q plot}
}
\description{
Produces a Q-Q plot of residuals from a linear model to test for normality, optionally annotating outlier points and adding the Shapiro-Wilk test p-value.
}
\details{
The plot visualizes the distribution of residuals against theoretical normal quantiles. Points are colored and shaped by outlier status, and a reference Q-Q line is added. Optionally, outlier and regular points can be labeled. If enabled, the Shapiro-Wilk normality test p-value is displayed.
}
\examples{
mdl <- lm(Sepal.Length ~ Sepal.Width, data = iris)
result <- lm_plot.qq(mdl)
print(result$plts$qq)
}
\seealso{
\code{\link{lm_plot.df}}, \code{\link{lm_plot.parms}}, \code{\link[stats]{shapiro.test}}
}
